+++ docs/_site/public/main.js
@@ -1,0 +1,12 @@
+/**
+ * Links to be displayed in the top right corner of the site.
+ */
+export default {
+  iconLinks: [
+    {
+      icon: 'github',
+      href: 'https://github.com/darwijesinghe/Snipster',
+      title: 'GitHub'
+    }
+  ]
+}